# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Swagger Doc For Find-Apps Api ",
        "version": "1.0.0",
        "title": "Restaurant Apis"
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "paths": {
...
        "/menu/getmenu": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " getmenu",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "getmenu ",
                "operationId": "getmenu",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getmenu"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/getmenu"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
...
        "/menu/getsection": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " getsection",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "getsection ",
                "operationId": "getsection",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getsection"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/getsection"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/getsectionbyid/{id}": {
            "get": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " getsectionbyid",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "getsectionbyid ",
                "operationId": "getsectionbyid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/getsectionbyid"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/getsectionbyid"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/updatesection/{id}": {
            "put": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " updatesection",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "updatesection ",
                "operationId": "updatesection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updatesection"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/updatesection"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/updatesection"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/changesectionstatus/{id}": {
            "put": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " changesectionstatus",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "changesectionstatus ",
                "operationId": "changesectionstatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/changesectionstatus"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/changesectionstatus"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        },
        "/menu/deletesection/{id}": {
            "delete": {
                "tags": [
                    "Restaurant"
                ],
                "summary": " deletesection",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "deletesection ",
                "operationId": "deletesection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/deletesection"
                            }
                        },
                        "application/xml": {
                            "schema": {
                                "$ref": "#/components/schemas/deletesection"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {}
            }
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:4000"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "in": "header",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "getmenu": {
                "type": "object",
                "required": [],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
...
            "getsection": {
                "type": "object",
                "required": [],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
            "getsectionbyid": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
            "updatesection": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "deserts"
                    },
                    "ar_name": {
                        "type": "string",
                        "example": "adsfg"
                    },
                    "menu_id": {
                        "type": "string",
                        "example": "45df5g4ger5gre54g4"
                    },
                    "image": {
                        "type": "file",
                        "example": "deserts.jpg"
                    },
                    "desc": {
                        "type": "string",
                        "example": "dfghqwertsd"
                    },
                    "ar_desc": {
                        "type": "string",
                        "example": "sdfghwertyuert"
                    }
                }
            },
            "changesectionstatus": {
                "type": "object",
                "required": [
                    "section_status"
                ],
                "properties": {
                    "section_status": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
            "deletesection": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    }
                }
            },
...
        }
    }
}